+++ dataforge-output/src/nativeMain/kotlin/hep/dataforge/output/outputNative.kt
@@ -1,6 +1,0 @@
-package space.kscience.dataforge.output
-
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-
-public actual val Dispatchers.Output: CoroutineDispatcher get() = Dispatchers.Default